requirements.txt
py-radix
arrow
confluent_kafka
msgpack
